refactor(query): Invoke the query interceptors from the query service
An interceptor was a tracestore.Reader decorator, which put it below the query
service and therefore after the query service had already chosen the shape
storage would receive. A filter query to a backend without filter support was
converted three times — to the legacy fields, back to a filter for the
interceptor, then to the legacy fields again — and the middle step is lossy,
because rewriting into the legacy fields drops the level off an attribute
predicate and rebuilding gives back an unqualified one. An interceptor deciding
policy by level therefore judged a predicate the caller never sent. Nothing
re-checked what the interceptor returned against the backend's declaration
either, so the contract documented a refusal that did not happen.

The interceptors are now a QueryService dependency, invoked between the checks
that do not depend on the backend and the capability call. The interceptor sees
the filter the caller sent, the query service converts once toward whatever the
backend declared, and that declaration covers the interceptor's own additions
because the check now runs after it.

Two paths gain behavior they should always have had: archived traces pass through
OnResult, which the decorator never saw because the archive reader was not
wrapped, and the trace-summary fallback runs OnQuery once rather than twice.
FindTraceIDs loses its hook, which gated nothing: no caller in jaeger-query
reaches the reader's FindTraceIDs, and a future one would be gated for free by
going through prepareSearchQuery.
