--- cmd/jaeger/internal/extension/jaegerquery/queryinterceptor/reader_decorator.go
@@ -1,251 +1,0 @@
-// Copyright (c) 2026 The Jaeger Authors.
-// SPDX-License-Identifier: Apache-2.0
-
-package queryinterceptor
-
-import (
-	"context"
-	"errors"
-	"fmt"
-	"iter"
-
-	"go.opentelemetry.io/collector/pdata/pcommon"
-	"go.opentelemetry.io/collector/pdata/ptrace"
-
-	expression "github.com/jaegertracing/jaeger-idl/query/expression/v1"
-	pub "github.com/jaegertracing/jaeger/components/extension/jaegerquery/queryinterceptor"
-	"github.com/jaegertracing/jaeger/internal/storage/v2/api/tracestore"
-)
-
-// ErrInterceptorFilter reports that a query interceptor returned a filter jaeger-query will not
-// send to storage. It is deliberately not one of the errors the API layers answer 400 for: the
-// caller's request was fine, and the fault is in the extension this deployment configured.
-var ErrInterceptorFilter = errors.New("query interceptor returned an invalid filter")
-
-type reader struct {
-	next         tracestore.Reader
-	interceptors []pub.Interceptor
-}
-
-// NewReaderDecorator decorates next so the given interceptors are applied around
-// every trace query: OnQuery on the query parameters of FindTraces,
-// FindTraceIDs, and FindTraceSummaries, and OnResult on every batch of traces
-// yielded by FindTraces and GetTraces. Interceptors run in the order given.
-// Callers wrap only when there is at least one interceptor, so this always
-// returns a decorator.
-//
-// The interceptors see the public queryinterceptor.Query; this decorator converts to and from
-// the internal tracestore.TraceQueryParams at the boundary, so the internal query type never
-// crosses the contract.
-func NewReaderDecorator(next tracestore.Reader, interceptors ...pub.Interceptor) tracestore.Reader {
-	return &reader{next: next, interceptors: interceptors}
-}
-
-func toPublicQuery(q tracestore.TraceQueryParams) pub.Query {
-	return pub.Query{
-		Filter:       q.Filter,
-		StartTimeMin: q.StartTimeMin,
-		StartTimeMax: q.StartTimeMax,
-		SearchDepth:  q.SearchDepth,
-	}
-}
-
-func fromPublicQuery(q pub.Query) tracestore.TraceQueryParams {
-	return tracestore.TraceQueryParams{
-		Attributes:   pcommon.NewMap(),
-		Filter:       q.Filter,
-		StartTimeMin: q.StartTimeMin,
-		StartTimeMax: q.StartTimeMax,
-		SearchDepth:  q.SearchDepth,
-	}
-}
-
-// onQuery runs every interceptor's OnQuery in order, threading the context each
-// returns into the next. The final context is returned so callers can pass it to
-// the storage reader and to onResult, letting an interceptor carry per-query state
-// (e.g. a resolved caller identity) from the pre-query hook to the return path.
-//
-// The interceptors are shown the query in filter shape whatever shape it arrived in, because
-// gating a search means reading and narrowing its predicates and an interceptor should not
-// have to find them in two places. What reaches storage afterwards is put back into the shape
-// that storage supports, which is the shape the query arrived in, because the query service
-// already chose it from what the reader declared: a reader that does not evaluate filters gets
-// the scalar fields, and one that does gets the filter the interceptors left behind. An
-// interceptor that adds a
-// predicate the scalar fields cannot express — a disjunction, say — is refused here rather than
-// silently dropped, which is the same answer the query service would have given for that filter.
-func (r *reader) onQuery(ctx context.Context, query tracestore.TraceQueryParams) (context.Context, tracestore.TraceQueryParams, error) {
-	wasFilterShaped := query.Filter != nil
-	asked := query.ToFilterShape()
-	pq := toPublicQuery(asked)
-	var err error
-	for _, interceptor := range r.interceptors {
-		ctx, pq, err = interceptor.OnQuery(ctx, pq)
-		if err != nil {
-			return ctx, query, err
-		}
-	}
-	if err := checkInterceptorFilter(pq.Filter, asked.Filter); err != nil {
-		return ctx, query, err
-	}
-	intercepted := fromPublicQuery(pq)
-	if wasFilterShaped {
-		return ctx, intercepted, nil
-	}
-	legacy, err := intercepted.ToLegacyShape()
-	if err != nil {
-		return ctx, query, err
-	}
-	return ctx, legacy, nil
-}
-
-// checkInterceptorFilter rejects what an interceptor must not hand to storage. An interceptor
-// builds its filter by hand, in code jaeger-query does not control, and a malformed tree is
-// typically answered by a backend matching nothing rather than refusing — so a search meant to
-// be narrowed would come back wrong with nothing to say why.
-//
-// Dropping the filter is refused separately, because it is the one mistake that fails open: a
-// search that arrived with predicates and leaves with none asks for every trace in the range.
-// It is only a mistake when there was something to drop, since a caller may legitimately search
-// a time range and nothing else.
-func checkInterceptorFilter(returned, asked *expression.Call) error {
-	if returned == nil {
-		if asked == nil {
-			return nil
-		}
-		return fmt.Errorf("%w: it returned no filter for a query that had predicates, which would "+
-			"widen the search to every trace in the time range", ErrInterceptorFilter)
-	}
-	if err := expression.ValidateFilter(returned); err != nil {
-		return fmt.Errorf("%w: %w", ErrInterceptorFilter, err)
-	}
-	return nil
-}
-
-// onResult runs every interceptor's OnResult in order on one batch, threading the
-// context each returns into the next. The final context is returned so the caller
-// can feed it into onResult for the next batch, letting state accumulate across a
-// multi-batch result.
-func (r *reader) onResult(ctx context.Context, traces []ptrace.Traces) (context.Context, []ptrace.Traces, error) {
-	var err error
-	for _, interceptor := range r.interceptors {
-		ctx, traces, err = interceptor.OnResult(ctx, traces)
-		if err != nil {
-			return ctx, nil, err
-		}
-	}
-	return ctx, traces, nil
-}
-
-func (r *reader) FindTraces(ctx context.Context, query tracestore.TraceQueryParams) iter.Seq2[[]ptrace.Traces, error] {
-	return func(yield func([]ptrace.Traces, error) bool) {
-		var err error
-		// onQuery returns the (possibly enriched) context and gated query; assign
-		// them back so every statement below sees the post-hook values, never the
-		// originals. The same applies to onResult's context inside the loop.
-		ctx, query, err = r.onQuery(ctx, query)
-		if err != nil {
-			yield(nil, err)
-			return
-		}
-		for traces, err := range r.next.FindTraces(ctx, query) {
-			if err != nil {
-				if !yield(nil, err) {
-					return
-				}
-				continue
-			}
-			ctx, traces, err = r.onResult(ctx, traces)
-			if err != nil {
-				// Per the Interceptor contract, an OnResult error aborts the stream:
-				// stop rather than emit further batches, which could leak results the
-				// failed sanitize/redaction was meant to withhold.
-				yield(nil, err)
-				return
-			}
-			if !yield(traces, nil) {
-				return
-			}
-		}
-	}
-}
-
-func (r *reader) FindTraceIDs(ctx context.Context, query tracestore.TraceQueryParams) iter.Seq2[[]tracestore.FoundTraceID, error] {
-	return func(yield func([]tracestore.FoundTraceID, error) bool) {
-		var err error
-		ctx, query, err = r.onQuery(ctx, query)
-		if err != nil {
-			yield(nil, err)
-			return
-		}
-		for ids, err := range r.next.FindTraceIDs(ctx, query) {
-			if !yield(ids, err) {
-				return
-			}
-		}
-	}
-}
-
-func (r *reader) GetTraces(ctx context.Context, traceIDs ...tracestore.GetTraceParams) iter.Seq2[[]ptrace.Traces, error] {
-	return func(yield func([]ptrace.Traces, error) bool) {
-		// GetTraces has no pre-query hook, so OnResult's context chain seeds from
-		// the inbound context; assigning it back threads across batches.
-		for traces, err := range r.next.GetTraces(ctx, traceIDs...) {
-			if err != nil {
-				if !yield(nil, err) {
-					return
-				}
-				continue
-			}
-			ctx, traces, err = r.onResult(ctx, traces)
-			if err != nil {
-				// Per the Interceptor contract, an OnResult error aborts the stream:
-				// stop rather than emit further batches, which could leak results the
-				// failed sanitize/redaction was meant to withhold.
-				yield(nil, err)
-				return
-			}
-			if !yield(traces, nil) {
-				return
-			}
-		}
-	}
-}
-
-// SearchCapabilities forwards the wrapped reader's declaration: an interceptor gates
-// and sanitizes queries, it does not change which query shapes the backend accepts.
-func (r *reader) SearchCapabilities(ctx context.Context) (tracestore.SearchCapabilities, error) {
-	return r.next.SearchCapabilities(ctx)
-}
-
-func (r *reader) GetServices(ctx context.Context) ([]string, error) {
-	return r.next.GetServices(ctx)
-}
-
-func (r *reader) GetOperations(ctx context.Context, query tracestore.OperationQueryParams) ([]tracestore.Operation, error) {
-	return r.next.GetOperations(ctx, query)
-}
-
-// FindTraceSummaries forwards to the wrapped reader's FindTraceSummaries,
-// applying OnQuery gating to the search. When the wrapped reader has no native
-// summary support it yields errors.ErrUnsupported and the query service falls
-// back to FindTraces (which this decorator also gates). OnResult is not applied
-// here: summaries carry only per-trace metadata (not spans), and the Interceptor
-// contract has no summary hook — so an interceptor gates/constrains the summary
-// search via OnQuery, but cannot drop or redact individual summaries. A dedicated
-// summary hook is left as a contract extension.
-func (r *reader) FindTraceSummaries(ctx context.Context, query tracestore.TraceQueryParams) iter.Seq2[[]tracestore.TraceSummary, error] {
-	return func(yield func([]tracestore.TraceSummary, error) bool) {
-		var err error
-		ctx, query, err = r.onQuery(ctx, query)
-		if err != nil {
-			yield(nil, err)
-			return
-		}
-		for summaries, err := range r.next.FindTraceSummaries(ctx, query) {
-			if !yield(summaries, err) {
-				return
-			}
-		}
-	}
-}
