+++ cmd/jaeger/internal/extension/jaegerquery/server.go
@@ -78,17 +78,13 @@ func (s *server) Start(ctx context.Context, host component.Host) error {
 		return fmt.Errorf("cannot create trace reader: %w", err)
 	}
 
-	// Resolve any configured query-interceptor extensions from the host and,
-	// if there are any, wrap the trace reader with them. This is the Option-D
-	// extension point: jaeger-query invokes a caller-supplied plugin (an OTel
-	// extension) around every trace query — gating the query and sanitizing the
-	// results — without exposing the storage Reader itself.
+	// Resolve any configured query-interceptor extensions from the host and hand them to the
+	// query service below. This is the Option-D extension point: jaeger-query invokes a
+	// caller-supplied plugin (an OTel extension) around every trace query — gating the query and
+	// sanitizing the results — without exposing the storage Reader itself.
 	interceptors, err := queryinterceptor.Resolve(host, s.config.QueryInterceptors)
 	if err != nil {
 		return fmt.Errorf("cannot resolve query interceptors: %w", err)
-	}
-	if len(interceptors) > 0 {
-		traceReader = queryinterceptor.NewReaderDecorator(traceReader, interceptors...)
 	}
 
 	df, ok := tf.(depstore.Factory)
@@ -103,6 +99,7 @@ func (s *server) Start(ctx context.Context, host component.Host) error {
 	opts := querysvc.QueryServiceOptions{
 		MaxClockSkewAdjust: s.config.MaxClockSkewAdjust,
 		MaxTraceSize:       s.config.MaxTraceSize,
+		Interceptors:       interceptors,
 	}
 	if err := s.addArchiveStorage(&opts, host); err != nil {
 		return err
