+++ docs/rfc/0005-structured-query-filters.md
@@ -644,7 +644,7 @@ PR-sized milestones with explicit exit bars, grouped into stages. The API is L2 
 **Stage A — API foundation (additive, no behavior change)**
 
 - **M1 — Proto types in jaeger-idl.** Add `Expression`, `Reference`, `Scalar`, `List`, and `Call` (with `level`/`op`/`type` as string enumerations whose closed sets are declared in the OpenAPI schema — §6.2) and the `filter` field on `TraceQueryParameters`, in both the public api_v3 and the storage/v2 protos; and the `filter`/`same_span_conjunction` capability fields on `storage.v2.SearchCapabilities` (§7). Legacy `attributes` untouched. *Initial delivery may ship the `ref` and `scalar` terms with span-level attributes and built-in fields, and phase in the `list` term, the non-span levels, and the `some` quantifier (§5.5), since the oneof and the `op` vocabulary are additive.* **In flight — [jaeger-idl#206](https://github.com/jaegertracing/jaeger-idl/pull/206), which encodes the recursive `Expression` + `Call` AST (the `ref`/`scalar`/`list`/`call` terms and the `level`/`op`/`type` string enumerations) per §6.1–§6.2, and the `FilterCapabilities` fields per §7.** *Exit:* generated types compile and vendor cleanly; existing api_v3 callers byte-for-byte unaffected.
-- **M2 — Plumb the filter through the query service to the storage interface.** Extend the internal `TraceQueryParams` ([`reader.go`](../../internal/storage/v2/api/tracestore/reader.go)) to carry the expression tree alongside the legacy `Attributes` map, and translate the proto field in the api_v3 handler. With no backend routing yet, a purely-conjunctive tree over indexed levels is treated as unqualified search-all (today's results); non-conjunctive trees, unsupported operators, and predicates naming a level the backend cannot honor (link, instrumentation on the flat backends; §1.6) are refused at the edge — up front where the query service can read the backend's declared filter capabilities (`SearchCapabilities`, [ADR-013](../adr/013-storage-capability-declaration.md)), at query time otherwise. *Exit:* a conjunctive filter over indexed levels reaches every backend as unqualified attributes and returns today's results; `OR`/`NOT`, unsupported operators, and unsupported levels are refused; plugins ignoring `filter` are unaffected.
+- **M2 ✅ — Plumb the filter through the query service to the storage interface** — delivered as a stack: [#9349](https://github.com/jaegertracing/jaeger/pull/9349) generated code, [#9354](https://github.com/jaegertracing/jaeger/pull/9354) the filter builder, [#9355](https://github.com/jaegertracing/jaeger/pull/9355) the wire conversion, [#9356](https://github.com/jaegertracing/jaeger/pull/9356) the storage declaration, [#9357](https://github.com/jaegertracing/jaeger/pull/9357) the query service, [#9358](https://github.com/jaegertracing/jaeger/pull/9358) remote storage, [#9359](https://github.com/jaegertracing/jaeger/pull/9359) the api_v3 edge, and [#9360](https://github.com/jaegertracing/jaeger/pull/9360) the query interceptor. Extend the internal `TraceQueryParams` ([`reader.go`](../../internal/storage/v2/api/tracestore/reader.go)) to carry the expression tree alongside the legacy `Attributes` map, and translate the proto field in the api_v3 handler. With no backend routing yet, a purely-conjunctive tree over indexed levels is treated as unqualified search-all (today's results); non-conjunctive trees, unsupported operators, and predicates naming a level the backend cannot honor (link, instrumentation on the flat backends; §1.6) are refused at the edge — up front where the query service can read the backend's declared filter capabilities (`SearchCapabilities`, [ADR-013](../adr/013-storage-capability-declaration.md)), at query time otherwise. *Exit:* a conjunctive filter over indexed levels reaches every backend as unqualified attributes and returns today's results; `OR`/`NOT`, unsupported operators, and unsupported levels are refused; plugins ignoring `filter` are unaffected. **As delivered it also carries** the feature gate and the interceptor contract described in §7, and the filter AST as Go types in jaeger-idl (`query/expression/v1`) — hand-written and independent of the generated protos, so that the public query API, the internal storage API and the interceptor contract all name one AST rather than converting between three spellings of it.
 
 **Stage B — Backend routing (one PR per backend, parallelizable after M2)**
